Type Definition tm4c123x::adc0::emux::EM0_R

pub type EM0_R = R<u8, EM0_A>;

Reader of field EM0

Implementations

impl EM0_R

pub fn variant(&self) -> Variant<u8, EM0_A>

Get enumerated values variant

pub fn is_processor(&self) -> bool

Checks if the value of the field is PROCESSOR

pub fn is_comp0(&self) -> bool

Checks if the value of the field is COMP0

pub fn is_comp1(&self) -> bool

Checks if the value of the field is COMP1

pub fn is_external(&self) -> bool

Checks if the value of the field is EXTERNAL

pub fn is_timer(&self) -> bool

Checks if the value of the field is TIMER

pub fn is_pwm0(&self) -> bool

Checks if the value of the field is PWM0

pub fn is_pwm1(&self) -> bool

Checks if the value of the field is PWM1

pub fn is_pwm2(&self) -> bool

Checks if the value of the field is PWM2

pub fn is_pwm3(&self) -> bool

Checks if the value of the field is PWM3

pub fn is_always(&self) -> bool

Checks if the value of the field is ALWAYS