Type Definition tm4c123x::adc0::spc::PHASE_R

pub type PHASE_R = R<u8, PHASE_A>;

Reader of field PHASE

Implementations

impl PHASE_R

pub fn variant(&self) -> PHASE_A

Get enumerated values variant

pub fn is_0(&self) -> bool

Checks if the value of the field is _0

pub fn is_22_5(&self) -> bool

Checks if the value of the field is _22_5

pub fn is_45(&self) -> bool

Checks if the value of the field is _45

pub fn is_67_5(&self) -> bool

Checks if the value of the field is _67_5

pub fn is_90(&self) -> bool

Checks if the value of the field is _90

pub fn is_112_5(&self) -> bool

Checks if the value of the field is _112_5

pub fn is_135(&self) -> bool

Checks if the value of the field is _135

pub fn is_157_5(&self) -> bool

Checks if the value of the field is _157_5

pub fn is_180(&self) -> bool

Checks if the value of the field is _180

pub fn is_202_5(&self) -> bool

Checks if the value of the field is _202_5

pub fn is_225(&self) -> bool

Checks if the value of the field is _225

pub fn is_247_5(&self) -> bool

Checks if the value of the field is _247_5

pub fn is_270(&self) -> bool

Checks if the value of the field is _270

pub fn is_292_5(&self) -> bool

Checks if the value of the field is _292_5

pub fn is_315(&self) -> bool

Checks if the value of the field is _315

pub fn is_337_5(&self) -> bool

Checks if the value of the field is _337_5