Type Definition tm4c123x::can0::sts::LEC_R

pub type LEC_R = R<u8, LEC_A>;

Reader of field LEC

Implementations

impl LEC_R

pub fn variant(&self) -> LEC_A

Get enumerated values variant

pub fn is_none(&self) -> bool

Checks if the value of the field is NONE

pub fn is_stuff(&self) -> bool

Checks if the value of the field is STUFF

pub fn is_form(&self) -> bool

Checks if the value of the field is FORM

pub fn is_ack(&self) -> bool

Checks if the value of the field is ACK

pub fn is_bit1(&self) -> bool

Checks if the value of the field is BIT1

pub fn is_bit0(&self) -> bool

Checks if the value of the field is BIT0

pub fn is_crc(&self) -> bool

Checks if the value of the field is CRC

pub fn is_noevent(&self) -> bool

Checks if the value of the field is NOEVENT