Type Definition tm4c123x::i2c0::mcr2::GFPW_R

pub type GFPW_R = R<u8, GFPW_A>;

Reader of field GFPW

Implementations

impl GFPW_R

pub fn variant(&self) -> GFPW_A

Get enumerated values variant

pub fn is_bypass(&self) -> bool

Checks if the value of the field is BYPASS

pub fn is_1(&self) -> bool

Checks if the value of the field is _1

pub fn is_2(&self) -> bool

Checks if the value of the field is _2

pub fn is_3(&self) -> bool

Checks if the value of the field is _3

pub fn is_4(&self) -> bool

Checks if the value of the field is _4

pub fn is_8(&self) -> bool

Checks if the value of the field is _8

pub fn is_16(&self) -> bool

Checks if the value of the field is _16

pub fn is_31(&self) -> bool

Checks if the value of the field is _31