Type Definition tm4c123x::pwm0::_0_ctl::GENAUPD_R

pub type GENAUPD_R = R<u8, GENAUPD_A>;

Reader of field GENAUPD

Implementations

impl GENAUPD_R

pub fn variant(&self) -> Variant<u8, GENAUPD_A>

Get enumerated values variant

pub fn is_i(&self) -> bool

Checks if the value of the field is I

pub fn is_ls(&self) -> bool

Checks if the value of the field is LS

pub fn is_gs(&self) -> bool

Checks if the value of the field is GS