Type Definition tm4c123x::pwm0::_0_gena::ACTZERO_R

pub type ACTZERO_R = R<u8, ACTZERO_A>;

Reader of field ACTZERO

Implementations

impl ACTZERO_R

pub fn variant(&self) -> ACTZERO_A

Get enumerated values variant

pub fn is_none(&self) -> bool

Checks if the value of the field is NONE

pub fn is_inv(&self) -> bool

Checks if the value of the field is INV

pub fn is_zero(&self) -> bool

Checks if the value of the field is ZERO

pub fn is_one(&self) -> bool

Checks if the value of the field is ONE