Type Definition tm4c123x::ssi0::cr0::DSS_R

pub type DSS_R = R<u8, DSS_A>;

Reader of field DSS

Implementations

impl DSS_R

pub fn variant(&self) -> Variant<u8, DSS_A>

Get enumerated values variant

pub fn is_4(&self) -> bool

Checks if the value of the field is _4

pub fn is_5(&self) -> bool

Checks if the value of the field is _5

pub fn is_6(&self) -> bool

Checks if the value of the field is _6

pub fn is_7(&self) -> bool

Checks if the value of the field is _7

pub fn is_8(&self) -> bool

Checks if the value of the field is _8

pub fn is_9(&self) -> bool

Checks if the value of the field is _9

pub fn is_10(&self) -> bool

Checks if the value of the field is _10

pub fn is_11(&self) -> bool

Checks if the value of the field is _11

pub fn is_12(&self) -> bool

Checks if the value of the field is _12

pub fn is_13(&self) -> bool

Checks if the value of the field is _13

pub fn is_14(&self) -> bool

Checks if the value of the field is _14

pub fn is_15(&self) -> bool

Checks if the value of the field is _15

pub fn is_16(&self) -> bool

Checks if the value of the field is _16