Module tm4c123x::sysctl

System Control register addresses

Modules

dc0

Device Capabilities 0

dc1

Device Capabilities 1

dc2

Device Capabilities 2

dc3

Device Capabilities 3

dc4

Device Capabilities 4

dc5

Device Capabilities 5

dc6

Device Capabilities 6

dc7

Device Capabilities 7

dc8

Device Capabilities 8

dc9

Device Capabilities 9

dcgc0

Deep Sleep Mode Clock Gating Control Register 0

dcgc1

Deep-Sleep Mode Clock Gating Control Register 1

dcgc2

Deep Sleep Mode Clock Gating Control Register 2

dcgcacmp

Analog Comparator Deep-Sleep Mode Clock Gating Control

dcgcadc

Analog-to-Digital Converter Deep-Sleep Mode Clock Gating Control

dcgccan

Controller Area Network Deep-Sleep Mode Clock Gating Control

dcgcdma

Micro Direct Memory Access Deep-Sleep Mode Clock Gating Control

dcgceeprom

EEPROM Deep-Sleep Mode Clock Gating Control

dcgcgpio

General-Purpose Input/Output Deep-Sleep Mode Clock Gating Control

dcgchib

Hibernation Deep-Sleep Mode Clock Gating Control

dcgci2c

Inter-Integrated Circuit Deep-Sleep Mode Clock Gating Control

dcgcpwm

Pulse Width Modulator Deep-Sleep Mode Clock Gating Control

dcgcqei

Quadrature Encoder Interface Deep-Sleep Mode Clock Gating Control

dcgcssi

Synchronous Serial Interface Deep-Sleep Mode Clock Gating Control

dcgctimer

16/32-Bit General-Purpose Timer Deep-Sleep Mode Clock Gating Control

dcgcuart

Universal Asynchronous Receiver/Transmitter Deep-Sleep Mode Clock Gating Control

dcgcusb

Universal Serial Bus Deep-Sleep Mode Clock Gating Control

dcgcwd

Watchdog Timer Deep-Sleep Mode Clock Gating Control

dcgcwtimer

32/64-Bit Wide General-Purpose Timer Deep-Sleep Mode Clock Gating Control

did0

Device Identification 0

did1

Device Identification 1

dslpclkcfg

Deep Sleep Clock Configuration

dslppwrcfg

Deep-Sleep Power Configuration

gpiohbctl

GPIO High-Performance Bus Control

imc

Interrupt Mask Control

ldodpctl

LDO Deep-Sleep Power Control

ldospctl

LDO Sleep Power Control

misc

Masked Interrupt Status and Clear

moscctl

Main Oscillator Control

nvmstat

Non-Volatile Memory Information

pborctl

Brown-Out Reset Control

piosccal

Precision Internal Oscillator Calibration

pioscstat

Precision Internal Oscillator Statistics

pllfreq0

PLL Frequency 0

pllfreq1

PLL Frequency 1

pllstat

PLL Status

ppacmp

Analog Comparator Peripheral Present

ppadc

Analog-to-Digital Converter Peripheral Present

ppcan

Controller Area Network Peripheral Present

ppdma

Micro Direct Memory Access Peripheral Present

ppeeprom

EEPROM Peripheral Present

ppgpio

General-Purpose Input/Output Peripheral Present

pphib

Hibernation Peripheral Present

ppi2c

Inter-Integrated Circuit Peripheral Present

pppwm

Pulse Width Modulator Peripheral Present

ppqei

Quadrature Encoder Interface Peripheral Present

ppssi

Synchronous Serial Interface Peripheral Present

pptimer

16/32-Bit General-Purpose Timer Peripheral Present

ppuart

Universal Asynchronous Receiver/Transmitter Peripheral Present

ppusb

Universal Serial Bus Peripheral Present

ppwd

Watchdog Timer Peripheral Present

ppwtimer

32/64-Bit Wide General-Purpose Timer Peripheral Present

pracmp

Analog Comparator Peripheral Ready

pradc

Analog-to-Digital Converter Peripheral Ready

prcan

Controller Area Network Peripheral Ready

prdma

Micro Direct Memory Access Peripheral Ready

preeprom

EEPROM Peripheral Ready

prgpio

General-Purpose Input/Output Peripheral Ready

prhib

Hibernation Peripheral Ready

pri2c

Inter-Integrated Circuit Peripheral Ready

prpwm

Pulse Width Modulator Peripheral Ready

prqei

Quadrature Encoder Interface Peripheral Ready

prssi

Synchronous Serial Interface Peripheral Ready

prtimer

16/32-Bit General-Purpose Timer Peripheral Ready

pruart

Universal Asynchronous Receiver/Transmitter Peripheral Ready

prusb

Universal Serial Bus Peripheral Ready

prwd

Watchdog Timer Peripheral Ready

prwtimer

32/64-Bit Wide General-Purpose Timer Peripheral Ready

rcc

Run-Mode Clock Configuration

rcc2

Run-Mode Clock Configuration 2

rcgc0

Run Mode Clock Gating Control Register 0

rcgc1

Run Mode Clock Gating Control Register 1

rcgc2

Run Mode Clock Gating Control Register 2

rcgcacmp

Analog Comparator Run Mode Clock Gating Control

rcgcadc

Analog-to-Digital Converter Run Mode Clock Gating Control

rcgccan

Controller Area Network Run Mode Clock Gating Control

rcgcdma

Micro Direct Memory Access Run Mode Clock Gating Control

rcgceeprom

EEPROM Run Mode Clock Gating Control

rcgcgpio

General-Purpose Input/Output Run Mode Clock Gating Control

rcgchib

Hibernation Run Mode Clock Gating Control

rcgci2c

Inter-Integrated Circuit Run Mode Clock Gating Control

rcgcpwm

Pulse Width Modulator Run Mode Clock Gating Control

rcgcqei

Quadrature Encoder Interface Run Mode Clock Gating Control

rcgcssi

Synchronous Serial Interface Run Mode Clock Gating Control

rcgctimer

16/32-Bit General-Purpose Timer Run Mode Clock Gating Control

rcgcuart

Universal Asynchronous Receiver/Transmitter Run Mode Clock Gating Control

rcgcusb

Universal Serial Bus Run Mode Clock Gating Control

rcgcwd

Watchdog Timer Run Mode Clock Gating Control

rcgcwtimer

32/64-Bit Wide General-Purpose Timer Run Mode Clock Gating Control

resc

Reset Cause

ris

Raw Interrupt Status

scgc0

Sleep Mode Clock Gating Control Register 0

scgc1

Sleep Mode Clock Gating Control Register 1

scgc2

Sleep Mode Clock Gating Control Register 2

scgcacmp

Analog Comparator Sleep Mode Clock Gating Control

scgcadc

Analog-to-Digital Converter Sleep Mode Clock Gating Control

scgccan

Controller Area Network Sleep Mode Clock Gating Control

scgcdma

Micro Direct Memory Access Sleep Mode Clock Gating Control

scgceeprom

EEPROM Sleep Mode Clock Gating Control

scgcgpio

General-Purpose Input/Output Sleep Mode Clock Gating Control

scgchib

Hibernation Sleep Mode Clock Gating Control

scgci2c

Inter-Integrated Circuit Sleep Mode Clock Gating Control

scgcpwm

Pulse Width Modulator Sleep Mode Clock Gating Control

scgcqei

Quadrature Encoder Interface Sleep Mode Clock Gating Control

scgcssi

Synchronous Serial Interface Sleep Mode Clock Gating Control

scgctimer

16/32-Bit General-Purpose Timer Sleep Mode Clock Gating Control

scgcuart

Universal Asynchronous Receiver/Transmitter Sleep Mode Clock Gating Control

scgcusb

Universal Serial Bus Sleep Mode Clock Gating Control

scgcwd

Watchdog Timer Sleep Mode Clock Gating Control

scgcwtimer

32/64-Bit Wide General-Purpose Timer Sleep Mode Clock Gating Control

slppwrcfg

Sleep Power Configuration

sracmp

Analog Comparator Software Reset

sradc

Analog-to-Digital Converter Software Reset

srcan

Controller Area Network Software Reset

srcr0

Software Reset Control 0

srcr1

Software Reset Control 1

srcr2

Software Reset Control 2

srdma

Micro Direct Memory Access Software Reset

sreeprom

EEPROM Software Reset

srgpio

General-Purpose Input/Output Software Reset

srhib

Hibernation Software Reset

sri2c

Inter-Integrated Circuit Software Reset

srpwm

Pulse Width Modulator Software Reset

srqei

Quadrature Encoder Interface Software Reset

srssi

Synchronous Serial Interface Software Reset

srtimer

16/32-Bit General-Purpose Timer Software Reset

sruart

Universal Asynchronous Receiver/Transmitter Software Reset

srusb

Universal Serial Bus Software Reset

srwd

Watchdog Timer Software Reset

srwtimer

32/64-Bit Wide General-Purpose Timer Software Reset

sysprop

System Properties

Structs

RegisterBlock

Register block

Type Definitions

DC0

Device Capabilities 0

DC1

Device Capabilities 1

DC2

Device Capabilities 2

DC3

Device Capabilities 3

DC4

Device Capabilities 4

DC5

Device Capabilities 5

DC6

Device Capabilities 6

DC7

Device Capabilities 7

DC8

Device Capabilities 8

DC9

Device Capabilities 9

DCGC0

Deep Sleep Mode Clock Gating Control Register 0

DCGC1

Deep-Sleep Mode Clock Gating Control Register 1

DCGC2

Deep Sleep Mode Clock Gating Control Register 2

DCGCACMP

Analog Comparator Deep-Sleep Mode Clock Gating Control

DCGCADC

Analog-to-Digital Converter Deep-Sleep Mode Clock Gating Control

DCGCCAN

Controller Area Network Deep-Sleep Mode Clock Gating Control

DCGCDMA

Micro Direct Memory Access Deep-Sleep Mode Clock Gating Control

DCGCEEPROM

EEPROM Deep-Sleep Mode Clock Gating Control

DCGCGPIO

General-Purpose Input/Output Deep-Sleep Mode Clock Gating Control

DCGCHIB

Hibernation Deep-Sleep Mode Clock Gating Control

DCGCI2C

Inter-Integrated Circuit Deep-Sleep Mode Clock Gating Control

DCGCPWM

Pulse Width Modulator Deep-Sleep Mode Clock Gating Control

DCGCQEI

Quadrature Encoder Interface Deep-Sleep Mode Clock Gating Control

DCGCSSI

Synchronous Serial Interface Deep-Sleep Mode Clock Gating Control

DCGCTIMER

16/32-Bit General-Purpose Timer Deep-Sleep Mode Clock Gating Control

DCGCUART

Universal Asynchronous Receiver/Transmitter Deep-Sleep Mode Clock Gating Control

DCGCUSB

Universal Serial Bus Deep-Sleep Mode Clock Gating Control

DCGCWD

Watchdog Timer Deep-Sleep Mode Clock Gating Control

DCGCWTIMER

32/64-Bit Wide General-Purpose Timer Deep-Sleep Mode Clock Gating Control

DID0

Device Identification 0

DID1

Device Identification 1

DSLPCLKCFG

Deep Sleep Clock Configuration

DSLPPWRCFG

Deep-Sleep Power Configuration

GPIOHBCTL

GPIO High-Performance Bus Control

IMC

Interrupt Mask Control

LDODPCTL

LDO Deep-Sleep Power Control

LDOSPCTL

LDO Sleep Power Control

MISC

Masked Interrupt Status and Clear

MOSCCTL

Main Oscillator Control

NVMSTAT

Non-Volatile Memory Information

PBORCTL

Brown-Out Reset Control

PIOSCCAL

Precision Internal Oscillator Calibration

PIOSCSTAT

Precision Internal Oscillator Statistics

PLLFREQ0

PLL Frequency 0

PLLFREQ1

PLL Frequency 1

PLLSTAT

PLL Status

PPACMP

Analog Comparator Peripheral Present

PPADC

Analog-to-Digital Converter Peripheral Present

PPCAN

Controller Area Network Peripheral Present

PPDMA

Micro Direct Memory Access Peripheral Present

PPEEPROM

EEPROM Peripheral Present

PPGPIO

General-Purpose Input/Output Peripheral Present

PPHIB

Hibernation Peripheral Present

PPI2C

Inter-Integrated Circuit Peripheral Present

PPPWM

Pulse Width Modulator Peripheral Present

PPQEI

Quadrature Encoder Interface Peripheral Present

PPSSI

Synchronous Serial Interface Peripheral Present

PPTIMER

16/32-Bit General-Purpose Timer Peripheral Present

PPUART

Universal Asynchronous Receiver/Transmitter Peripheral Present

PPUSB

Universal Serial Bus Peripheral Present

PPWD

Watchdog Timer Peripheral Present

PPWTIMER

32/64-Bit Wide General-Purpose Timer Peripheral Present

PRACMP

Analog Comparator Peripheral Ready

PRADC

Analog-to-Digital Converter Peripheral Ready

PRCAN

Controller Area Network Peripheral Ready

PRDMA

Micro Direct Memory Access Peripheral Ready

PREEPROM

EEPROM Peripheral Ready

PRGPIO

General-Purpose Input/Output Peripheral Ready

PRHIB

Hibernation Peripheral Ready

PRI2C

Inter-Integrated Circuit Peripheral Ready

PRPWM

Pulse Width Modulator Peripheral Ready

PRQEI

Quadrature Encoder Interface Peripheral Ready

PRSSI

Synchronous Serial Interface Peripheral Ready

PRTIMER

16/32-Bit General-Purpose Timer Peripheral Ready

PRUART

Universal Asynchronous Receiver/Transmitter Peripheral Ready

PRUSB

Universal Serial Bus Peripheral Ready

PRWD

Watchdog Timer Peripheral Ready

PRWTIMER

32/64-Bit Wide General-Purpose Timer Peripheral Ready

RCC

Run-Mode Clock Configuration

RCC2

Run-Mode Clock Configuration 2

RCGC0

Run Mode Clock Gating Control Register 0

RCGC1

Run Mode Clock Gating Control Register 1

RCGC2

Run Mode Clock Gating Control Register 2

RCGCACMP

Analog Comparator Run Mode Clock Gating Control

RCGCADC

Analog-to-Digital Converter Run Mode Clock Gating Control

RCGCCAN

Controller Area Network Run Mode Clock Gating Control

RCGCDMA

Micro Direct Memory Access Run Mode Clock Gating Control

RCGCEEPROM

EEPROM Run Mode Clock Gating Control

RCGCGPIO

General-Purpose Input/Output Run Mode Clock Gating Control

RCGCHIB

Hibernation Run Mode Clock Gating Control

RCGCI2C

Inter-Integrated Circuit Run Mode Clock Gating Control

RCGCPWM

Pulse Width Modulator Run Mode Clock Gating Control

RCGCQEI

Quadrature Encoder Interface Run Mode Clock Gating Control

RCGCSSI

Synchronous Serial Interface Run Mode Clock Gating Control

RCGCTIMER

16/32-Bit General-Purpose Timer Run Mode Clock Gating Control

RCGCUART

Universal Asynchronous Receiver/Transmitter Run Mode Clock Gating Control

RCGCUSB

Universal Serial Bus Run Mode Clock Gating Control

RCGCWD

Watchdog Timer Run Mode Clock Gating Control

RCGCWTIMER

32/64-Bit Wide General-Purpose Timer Run Mode Clock Gating Control

RESC

Reset Cause

RIS

Raw Interrupt Status

SCGC0

Sleep Mode Clock Gating Control Register 0

SCGC1

Sleep Mode Clock Gating Control Register 1

SCGC2

Sleep Mode Clock Gating Control Register 2

SCGCACMP

Analog Comparator Sleep Mode Clock Gating Control

SCGCADC

Analog-to-Digital Converter Sleep Mode Clock Gating Control

SCGCCAN

Controller Area Network Sleep Mode Clock Gating Control

SCGCDMA

Micro Direct Memory Access Sleep Mode Clock Gating Control

SCGCEEPROM

EEPROM Sleep Mode Clock Gating Control

SCGCGPIO

General-Purpose Input/Output Sleep Mode Clock Gating Control

SCGCHIB

Hibernation Sleep Mode Clock Gating Control

SCGCI2C

Inter-Integrated Circuit Sleep Mode Clock Gating Control

SCGCPWM

Pulse Width Modulator Sleep Mode Clock Gating Control

SCGCQEI

Quadrature Encoder Interface Sleep Mode Clock Gating Control

SCGCSSI

Synchronous Serial Interface Sleep Mode Clock Gating Control

SCGCTIMER

16/32-Bit General-Purpose Timer Sleep Mode Clock Gating Control

SCGCUART

Universal Asynchronous Receiver/Transmitter Sleep Mode Clock Gating Control

SCGCUSB

Universal Serial Bus Sleep Mode Clock Gating Control

SCGCWD

Watchdog Timer Sleep Mode Clock Gating Control

SCGCWTIMER

32/64-Bit Wide General-Purpose Timer Sleep Mode Clock Gating Control

SLPPWRCFG

Sleep Power Configuration

SRACMP

Analog Comparator Software Reset

SRADC

Analog-to-Digital Converter Software Reset

SRCAN

Controller Area Network Software Reset

SRCR0

Software Reset Control 0

SRCR1

Software Reset Control 1

SRCR2

Software Reset Control 2

SRDMA

Micro Direct Memory Access Software Reset

SREEPROM

EEPROM Software Reset

SRGPIO

General-Purpose Input/Output Software Reset

SRHIB

Hibernation Software Reset

SRI2C

Inter-Integrated Circuit Software Reset

SRPWM

Pulse Width Modulator Software Reset

SRQEI

Quadrature Encoder Interface Software Reset

SRSSI

Synchronous Serial Interface Software Reset

SRTIMER

16/32-Bit General-Purpose Timer Software Reset

SRUART

Universal Asynchronous Receiver/Transmitter Software Reset

SRUSB

Universal Serial Bus Software Reset

SRWD

Watchdog Timer Software Reset

SRWTIMER

32/64-Bit Wide General-Purpose Timer Software Reset

SYSPROP

System Properties