Type Definition tm4c123x::sysctl::dc0::FLASHSZ_R

pub type FLASHSZ_R = R<u16, FLASHSZ_A>;

Reader of field FLASHSZ

Implementations

impl FLASHSZ_R

pub fn variant(&self) -> Variant<u16, FLASHSZ_A>

Get enumerated values variant

pub fn is_8kb(&self) -> bool

Checks if the value of the field is _8KB

pub fn is_16kb(&self) -> bool

Checks if the value of the field is _16KB

pub fn is_32kb(&self) -> bool

Checks if the value of the field is _32KB

pub fn is_64kb(&self) -> bool

Checks if the value of the field is _64KB

pub fn is_96kb(&self) -> bool

Checks if the value of the field is _96KB

pub fn is_128k(&self) -> bool

Checks if the value of the field is _128K

pub fn is_192k(&self) -> bool

Checks if the value of the field is _192K

pub fn is_256k(&self) -> bool

Checks if the value of the field is _256K