Type Definition tm4c123x::sysctl::dc0::SRAMSZ_R

pub type SRAMSZ_R = R<u16, SRAMSZ_A>;

Reader of field SRAMSZ

Implementations

impl SRAMSZ_R

pub fn variant(&self) -> Variant<u16, SRAMSZ_A>

Get enumerated values variant

pub fn is_2kb(&self) -> bool

Checks if the value of the field is _2KB

pub fn is_4kb(&self) -> bool

Checks if the value of the field is _4KB

pub fn is_6kb(&self) -> bool

Checks if the value of the field is _6KB

pub fn is_8kb(&self) -> bool

Checks if the value of the field is _8KB

pub fn is_12kb(&self) -> bool

Checks if the value of the field is _12KB

pub fn is_16kb(&self) -> bool

Checks if the value of the field is _16KB

pub fn is_20kb(&self) -> bool

Checks if the value of the field is _20KB

pub fn is_24kb(&self) -> bool

Checks if the value of the field is _24KB

pub fn is_32kb(&self) -> bool

Checks if the value of the field is _32KB