Type Definition tm4c123x::sysctl::ldospctl::VLDO_R

pub type VLDO_R = R<u8, VLDO_A>;

Reader of field VLDO

Implementations

impl VLDO_R

pub fn variant(&self) -> Variant<u8, VLDO_A>

Get enumerated values variant

pub fn is_0_90v(&self) -> bool

Checks if the value of the field is _0_90V

pub fn is_0_95v(&self) -> bool

Checks if the value of the field is _0_95V

pub fn is_1_00v(&self) -> bool

Checks if the value of the field is _1_00V

pub fn is_1_05v(&self) -> bool

Checks if the value of the field is _1_05V

pub fn is_1_10v(&self) -> bool

Checks if the value of the field is _1_10V

pub fn is_1_15v(&self) -> bool

Checks if the value of the field is _1_15V

pub fn is_1_20v(&self) -> bool

Checks if the value of the field is _1_20V