Type Definition tm4c123x::sysctl::rcc::XTAL_R

pub type XTAL_R = R<u8, XTAL_A>;

Reader of field XTAL

Implementations

impl XTAL_R

pub fn variant(&self) -> Variant<u8, XTAL_A>

Get enumerated values variant

pub fn is_4mhz(&self) -> bool

Checks if the value of the field is _4MHZ

pub fn is_4_09mhz(&self) -> bool

Checks if the value of the field is _4_09MHZ

pub fn is_4_91mhz(&self) -> bool

Checks if the value of the field is _4_91MHZ

pub fn is_5mhz(&self) -> bool

Checks if the value of the field is _5MHZ

pub fn is_5_12mhz(&self) -> bool

Checks if the value of the field is _5_12MHZ

pub fn is_6mhz(&self) -> bool

Checks if the value of the field is _6MHZ

pub fn is_6_14mhz(&self) -> bool

Checks if the value of the field is _6_14MHZ

pub fn is_7_37mhz(&self) -> bool

Checks if the value of the field is _7_37MHZ

pub fn is_8mhz(&self) -> bool

Checks if the value of the field is _8MHZ

pub fn is_8_19mhz(&self) -> bool

Checks if the value of the field is _8_19MHZ

pub fn is_10mhz(&self) -> bool

Checks if the value of the field is _10MHZ

pub fn is_12mhz(&self) -> bool

Checks if the value of the field is _12MHZ

pub fn is_12_2mhz(&self) -> bool

Checks if the value of the field is _12_2MHZ

pub fn is_13_5mhz(&self) -> bool

Checks if the value of the field is _13_5MHZ

pub fn is_14_3mhz(&self) -> bool

Checks if the value of the field is _14_3MHZ

pub fn is_16mhz(&self) -> bool

Checks if the value of the field is _16MHZ

pub fn is_16_3mhz(&self) -> bool

Checks if the value of the field is _16_3MHZ

pub fn is_18mhz(&self) -> bool

Checks if the value of the field is _18MHZ

pub fn is_20mhz(&self) -> bool

Checks if the value of the field is _20MHZ

pub fn is_24mhz(&self) -> bool

Checks if the value of the field is _24MHZ

pub fn is_25mhz(&self) -> bool

Checks if the value of the field is _25MHZ