Struct tm4c123x::sysctl::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub did0: DID0,
    pub did1: DID1,
    pub dc0: DC0,
    pub dc1: DC1,
    pub dc2: DC2,
    pub dc3: DC3,
    pub dc4: DC4,
    pub dc5: DC5,
    pub dc6: DC6,
    pub dc7: DC7,
    pub dc8: DC8,
    pub pborctl: PBORCTL,
    pub srcr0: SRCR0,
    pub srcr1: SRCR1,
    pub srcr2: SRCR2,
    pub ris: RIS,
    pub imc: IMC,
    pub misc: MISC,
    pub resc: RESC,
    pub rcc: RCC,
    pub gpiohbctl: GPIOHBCTL,
    pub rcc2: RCC2,
    pub moscctl: MOSCCTL,
    pub rcgc0: RCGC0,
    pub rcgc1: RCGC1,
    pub rcgc2: RCGC2,
    pub scgc0: SCGC0,
    pub scgc1: SCGC1,
    pub scgc2: SCGC2,
    pub dcgc0: DCGC0,
    pub dcgc1: DCGC1,
    pub dcgc2: DCGC2,
    pub dslpclkcfg: DSLPCLKCFG,
    pub sysprop: SYSPROP,
    pub piosccal: PIOSCCAL,
    pub pioscstat: PIOSCSTAT,
    pub pllfreq0: PLLFREQ0,
    pub pllfreq1: PLLFREQ1,
    pub pllstat: PLLSTAT,
    pub slppwrcfg: SLPPWRCFG,
    pub dslppwrcfg: DSLPPWRCFG,
    pub dc9: DC9,
    pub nvmstat: NVMSTAT,
    pub ldospctl: LDOSPCTL,
    pub ldodpctl: LDODPCTL,
    pub ppwd: PPWD,
    pub pptimer: PPTIMER,
    pub ppgpio: PPGPIO,
    pub ppdma: PPDMA,
    pub pphib: PPHIB,
    pub ppuart: PPUART,
    pub ppssi: PPSSI,
    pub ppi2c: PPI2C,
    pub ppusb: PPUSB,
    pub ppcan: PPCAN,
    pub ppadc: PPADC,
    pub ppacmp: PPACMP,
    pub pppwm: PPPWM,
    pub ppqei: PPQEI,
    pub ppeeprom: PPEEPROM,
    pub ppwtimer: PPWTIMER,
    pub srwd: SRWD,
    pub srtimer: SRTIMER,
    pub srgpio: SRGPIO,
    pub srdma: SRDMA,
    pub srhib: SRHIB,
    pub sruart: SRUART,
    pub srssi: SRSSI,
    pub sri2c: SRI2C,
    pub srusb: SRUSB,
    pub srcan: SRCAN,
    pub sradc: SRADC,
    pub sracmp: SRACMP,
    pub srpwm: SRPWM,
    pub srqei: SRQEI,
    pub sreeprom: SREEPROM,
    pub srwtimer: SRWTIMER,
    pub rcgcwd: RCGCWD,
    pub rcgctimer: RCGCTIMER,
    pub rcgcgpio: RCGCGPIO,
    pub rcgcdma: RCGCDMA,
    pub rcgchib: RCGCHIB,
    pub rcgcuart: RCGCUART,
    pub rcgcssi: RCGCSSI,
    pub rcgci2c: RCGCI2C,
    pub rcgcusb: RCGCUSB,
    pub rcgccan: RCGCCAN,
    pub rcgcadc: RCGCADC,
    pub rcgcacmp: RCGCACMP,
    pub rcgcpwm: RCGCPWM,
    pub rcgcqei: RCGCQEI,
    pub rcgceeprom: RCGCEEPROM,
    pub rcgcwtimer: RCGCWTIMER,
    pub scgcwd: SCGCWD,
    pub scgctimer: SCGCTIMER,
    pub scgcgpio: SCGCGPIO,
    pub scgcdma: SCGCDMA,
    pub scgchib: SCGCHIB,
    pub scgcuart: SCGCUART,
    pub scgcssi: SCGCSSI,
    pub scgci2c: SCGCI2C,
    pub scgcusb: SCGCUSB,
    pub scgccan: SCGCCAN,
    pub scgcadc: SCGCADC,
    pub scgcacmp: SCGCACMP,
    pub scgcpwm: SCGCPWM,
    pub scgcqei: SCGCQEI,
    pub scgceeprom: SCGCEEPROM,
    pub scgcwtimer: SCGCWTIMER,
    pub dcgcwd: DCGCWD,
    pub dcgctimer: DCGCTIMER,
    pub dcgcgpio: DCGCGPIO,
    pub dcgcdma: DCGCDMA,
    pub dcgchib: DCGCHIB,
    pub dcgcuart: DCGCUART,
    pub dcgcssi: DCGCSSI,
    pub dcgci2c: DCGCI2C,
    pub dcgcusb: DCGCUSB,
    pub dcgccan: DCGCCAN,
    pub dcgcadc: DCGCADC,
    pub dcgcacmp: DCGCACMP,
    pub dcgcpwm: DCGCPWM,
    pub dcgcqei: DCGCQEI,
    pub dcgceeprom: DCGCEEPROM,
    pub dcgcwtimer: DCGCWTIMER,
    pub prwd: PRWD,
    pub prtimer: PRTIMER,
    pub prgpio: PRGPIO,
    pub prdma: PRDMA,
    pub prhib: PRHIB,
    pub pruart: PRUART,
    pub prssi: PRSSI,
    pub pri2c: PRI2C,
    pub prusb: PRUSB,
    pub prcan: PRCAN,
    pub pradc: PRADC,
    pub pracmp: PRACMP,
    pub prpwm: PRPWM,
    pub prqei: PRQEI,
    pub preeprom: PREEPROM,
    pub prwtimer: PRWTIMER,
    /* private fields */

}

Register block

Fields
did0: DID0

0x00 - Device Identification 0

did1: DID1

0x04 - Device Identification 1

dc0: DC0

0x08 - Device Capabilities 0

dc1: DC1

0x10 - Device Capabilities 1

dc2: DC2

0x14 - Device Capabilities 2

dc3: DC3

0x18 - Device Capabilities 3

dc4: DC4

0x1c - Device Capabilities 4

dc5: DC5

0x20 - Device Capabilities 5

dc6: DC6

0x24 - Device Capabilities 6

dc7: DC7

0x28 - Device Capabilities 7

dc8: DC8

0x2c - Device Capabilities 8

pborctl: PBORCTL

0x30 - Brown-Out Reset Control

srcr0: SRCR0

0x40 - Software Reset Control 0

srcr1: SRCR1

0x44 - Software Reset Control 1

srcr2: SRCR2

0x48 - Software Reset Control 2

ris: RIS

0x50 - Raw Interrupt Status

imc: IMC

0x54 - Interrupt Mask Control

misc: MISC

0x58 - Masked Interrupt Status and Clear

resc: RESC

0x5c - Reset Cause

rcc: RCC

0x60 - Run-Mode Clock Configuration

gpiohbctl: GPIOHBCTL

0x6c - GPIO High-Performance Bus Control

rcc2: RCC2

0x70 - Run-Mode Clock Configuration 2

moscctl: MOSCCTL

0x7c - Main Oscillator Control

rcgc0: RCGC0

0x100 - Run Mode Clock Gating Control Register 0

rcgc1: RCGC1

0x104 - Run Mode Clock Gating Control Register 1

rcgc2: RCGC2

0x108 - Run Mode Clock Gating Control Register 2

scgc0: SCGC0

0x110 - Sleep Mode Clock Gating Control Register 0

scgc1: SCGC1

0x114 - Sleep Mode Clock Gating Control Register 1

scgc2: SCGC2

0x118 - Sleep Mode Clock Gating Control Register 2

dcgc0: DCGC0

0x120 - Deep Sleep Mode Clock Gating Control Register 0

dcgc1: DCGC1

0x124 - Deep-Sleep Mode Clock Gating Control Register 1

dcgc2: DCGC2

0x128 - Deep Sleep Mode Clock Gating Control Register 2

dslpclkcfg: DSLPCLKCFG

0x144 - Deep Sleep Clock Configuration

sysprop: SYSPROP

0x14c - System Properties

piosccal: PIOSCCAL

0x150 - Precision Internal Oscillator Calibration

pioscstat: PIOSCSTAT

0x154 - Precision Internal Oscillator Statistics

pllfreq0: PLLFREQ0

0x160 - PLL Frequency 0

pllfreq1: PLLFREQ1

0x164 - PLL Frequency 1

pllstat: PLLSTAT

0x168 - PLL Status

slppwrcfg: SLPPWRCFG

0x188 - Sleep Power Configuration

dslppwrcfg: DSLPPWRCFG

0x18c - Deep-Sleep Power Configuration

dc9: DC9

0x190 - Device Capabilities 9

nvmstat: NVMSTAT

0x1a0 - Non-Volatile Memory Information

ldospctl: LDOSPCTL

0x1b4 - LDO Sleep Power Control

ldodpctl: LDODPCTL

0x1bc - LDO Deep-Sleep Power Control

ppwd: PPWD

0x300 - Watchdog Timer Peripheral Present

pptimer: PPTIMER

0x304 - 16/32-Bit General-Purpose Timer Peripheral Present

ppgpio: PPGPIO

0x308 - General-Purpose Input/Output Peripheral Present

ppdma: PPDMA

0x30c - Micro Direct Memory Access Peripheral Present

pphib: PPHIB

0x314 - Hibernation Peripheral Present

ppuart: PPUART

0x318 - Universal Asynchronous Receiver/Transmitter Peripheral Present

ppssi: PPSSI

0x31c - Synchronous Serial Interface Peripheral Present

ppi2c: PPI2C

0x320 - Inter-Integrated Circuit Peripheral Present

ppusb: PPUSB

0x328 - Universal Serial Bus Peripheral Present

ppcan: PPCAN

0x334 - Controller Area Network Peripheral Present

ppadc: PPADC

0x338 - Analog-to-Digital Converter Peripheral Present

ppacmp: PPACMP

0x33c - Analog Comparator Peripheral Present

pppwm: PPPWM

0x340 - Pulse Width Modulator Peripheral Present

ppqei: PPQEI

0x344 - Quadrature Encoder Interface Peripheral Present

ppeeprom: PPEEPROM

0x358 - EEPROM Peripheral Present

ppwtimer: PPWTIMER

0x35c - 32/64-Bit Wide General-Purpose Timer Peripheral Present

srwd: SRWD

0x500 - Watchdog Timer Software Reset

srtimer: SRTIMER

0x504 - 16/32-Bit General-Purpose Timer Software Reset

srgpio: SRGPIO

0x508 - General-Purpose Input/Output Software Reset

srdma: SRDMA

0x50c - Micro Direct Memory Access Software Reset

srhib: SRHIB

0x514 - Hibernation Software Reset

sruart: SRUART

0x518 - Universal Asynchronous Receiver/Transmitter Software Reset

srssi: SRSSI

0x51c - Synchronous Serial Interface Software Reset

sri2c: SRI2C

0x520 - Inter-Integrated Circuit Software Reset

srusb: SRUSB

0x528 - Universal Serial Bus Software Reset

srcan: SRCAN

0x534 - Controller Area Network Software Reset

sradc: SRADC

0x538 - Analog-to-Digital Converter Software Reset

sracmp: SRACMP

0x53c - Analog Comparator Software Reset

srpwm: SRPWM

0x540 - Pulse Width Modulator Software Reset

srqei: SRQEI

0x544 - Quadrature Encoder Interface Software Reset

sreeprom: SREEPROM

0x558 - EEPROM Software Reset

srwtimer: SRWTIMER

0x55c - 32/64-Bit Wide General-Purpose Timer Software Reset

rcgcwd: RCGCWD

0x600 - Watchdog Timer Run Mode Clock Gating Control

rcgctimer: RCGCTIMER

0x604 - 16/32-Bit General-Purpose Timer Run Mode Clock Gating Control

rcgcgpio: RCGCGPIO

0x608 - General-Purpose Input/Output Run Mode Clock Gating Control

rcgcdma: RCGCDMA

0x60c - Micro Direct Memory Access Run Mode Clock Gating Control

rcgchib: RCGCHIB

0x614 - Hibernation Run Mode Clock Gating Control

rcgcuart: RCGCUART

0x618 - Universal Asynchronous Receiver/Transmitter Run Mode Clock Gating Control

rcgcssi: RCGCSSI

0x61c - Synchronous Serial Interface Run Mode Clock Gating Control

rcgci2c: RCGCI2C

0x620 - Inter-Integrated Circuit Run Mode Clock Gating Control

rcgcusb: RCGCUSB

0x628 - Universal Serial Bus Run Mode Clock Gating Control

rcgccan: RCGCCAN

0x634 - Controller Area Network Run Mode Clock Gating Control

rcgcadc: RCGCADC

0x638 - Analog-to-Digital Converter Run Mode Clock Gating Control

rcgcacmp: RCGCACMP

0x63c - Analog Comparator Run Mode Clock Gating Control

rcgcpwm: RCGCPWM

0x640 - Pulse Width Modulator Run Mode Clock Gating Control

rcgcqei: RCGCQEI

0x644 - Quadrature Encoder Interface Run Mode Clock Gating Control

rcgceeprom: RCGCEEPROM

0x658 - EEPROM Run Mode Clock Gating Control

rcgcwtimer: RCGCWTIMER

0x65c - 32/64-Bit Wide General-Purpose Timer Run Mode Clock Gating Control

scgcwd: SCGCWD

0x700 - Watchdog Timer Sleep Mode Clock Gating Control

scgctimer: SCGCTIMER

0x704 - 16/32-Bit General-Purpose Timer Sleep Mode Clock Gating Control

scgcgpio: SCGCGPIO

0x708 - General-Purpose Input/Output Sleep Mode Clock Gating Control

scgcdma: SCGCDMA

0x70c - Micro Direct Memory Access Sleep Mode Clock Gating Control

scgchib: SCGCHIB

0x714 - Hibernation Sleep Mode Clock Gating Control

scgcuart: SCGCUART

0x718 - Universal Asynchronous Receiver/Transmitter Sleep Mode Clock Gating Control

scgcssi: SCGCSSI

0x71c - Synchronous Serial Interface Sleep Mode Clock Gating Control

scgci2c: SCGCI2C

0x720 - Inter-Integrated Circuit Sleep Mode Clock Gating Control

scgcusb: SCGCUSB

0x728 - Universal Serial Bus Sleep Mode Clock Gating Control

scgccan: SCGCCAN

0x734 - Controller Area Network Sleep Mode Clock Gating Control

scgcadc: SCGCADC

0x738 - Analog-to-Digital Converter Sleep Mode Clock Gating Control

scgcacmp: SCGCACMP

0x73c - Analog Comparator Sleep Mode Clock Gating Control

scgcpwm: SCGCPWM

0x740 - Pulse Width Modulator Sleep Mode Clock Gating Control

scgcqei: SCGCQEI

0x744 - Quadrature Encoder Interface Sleep Mode Clock Gating Control

scgceeprom: SCGCEEPROM

0x758 - EEPROM Sleep Mode Clock Gating Control

scgcwtimer: SCGCWTIMER

0x75c - 32/64-Bit Wide General-Purpose Timer Sleep Mode Clock Gating Control

dcgcwd: DCGCWD

0x800 - Watchdog Timer Deep-Sleep Mode Clock Gating Control

dcgctimer: DCGCTIMER

0x804 - 16/32-Bit General-Purpose Timer Deep-Sleep Mode Clock Gating Control

dcgcgpio: DCGCGPIO

0x808 - General-Purpose Input/Output Deep-Sleep Mode Clock Gating Control

dcgcdma: DCGCDMA

0x80c - Micro Direct Memory Access Deep-Sleep Mode Clock Gating Control

dcgchib: DCGCHIB

0x814 - Hibernation Deep-Sleep Mode Clock Gating Control

dcgcuart: DCGCUART

0x818 - Universal Asynchronous Receiver/Transmitter Deep-Sleep Mode Clock Gating Control

dcgcssi: DCGCSSI

0x81c - Synchronous Serial Interface Deep-Sleep Mode Clock Gating Control

dcgci2c: DCGCI2C

0x820 - Inter-Integrated Circuit Deep-Sleep Mode Clock Gating Control

dcgcusb: DCGCUSB

0x828 - Universal Serial Bus Deep-Sleep Mode Clock Gating Control

dcgccan: DCGCCAN

0x834 - Controller Area Network Deep-Sleep Mode Clock Gating Control

dcgcadc: DCGCADC

0x838 - Analog-to-Digital Converter Deep-Sleep Mode Clock Gating Control

dcgcacmp: DCGCACMP

0x83c - Analog Comparator Deep-Sleep Mode Clock Gating Control

dcgcpwm: DCGCPWM

0x840 - Pulse Width Modulator Deep-Sleep Mode Clock Gating Control

dcgcqei: DCGCQEI

0x844 - Quadrature Encoder Interface Deep-Sleep Mode Clock Gating Control

dcgceeprom: DCGCEEPROM

0x858 - EEPROM Deep-Sleep Mode Clock Gating Control

dcgcwtimer: DCGCWTIMER

0x85c - 32/64-Bit Wide General-Purpose Timer Deep-Sleep Mode Clock Gating Control

prwd: PRWD

0xa00 - Watchdog Timer Peripheral Ready

prtimer: PRTIMER

0xa04 - 16/32-Bit General-Purpose Timer Peripheral Ready

prgpio: PRGPIO

0xa08 - General-Purpose Input/Output Peripheral Ready

prdma: PRDMA

0xa0c - Micro Direct Memory Access Peripheral Ready

prhib: PRHIB

0xa14 - Hibernation Peripheral Ready

pruart: PRUART

0xa18 - Universal Asynchronous Receiver/Transmitter Peripheral Ready

prssi: PRSSI

0xa1c - Synchronous Serial Interface Peripheral Ready

pri2c: PRI2C

0xa20 - Inter-Integrated Circuit Peripheral Ready

prusb: PRUSB

0xa28 - Universal Serial Bus Peripheral Ready

prcan: PRCAN

0xa34 - Controller Area Network Peripheral Ready

pradc: PRADC

0xa38 - Analog-to-Digital Converter Peripheral Ready

pracmp: PRACMP

0xa3c - Analog Comparator Peripheral Ready

prpwm: PRPWM

0xa40 - Pulse Width Modulator Peripheral Ready

prqei: PRQEI

0xa44 - Quadrature Encoder Interface Peripheral Ready

preeprom: PREEPROM

0xa58 - EEPROM Peripheral Ready

prwtimer: PRWTIMER

0xa5c - 32/64-Bit Wide General-Purpose Timer Peripheral Ready