Module tm4c123x::usb0

Univeral Serial Bus register offsets

Modules

contim

USB Connect Timing

count0

USB Receive Byte Count Endpoint 0

csrh0

USB Control and Status Endpoint 0 High

csrl0

USB Control and Status Endpoint 0 Low

devctl

USB Device Control

dmasel

USB DMA Select

drim

USB Device RESUME Interrupt Mask

drisc

USB Device RESUME Interrupt Status and Clear

drris

USB Device RESUME Raw Interrupt Status

epc

USB External Power Control

epcim

USB External Power Control Interrupt Mask

epcisc

USB External Power Control Interrupt Status and Clear

epcris

USB External Power Control Raw Interrupt Status

epidx

USB Endpoint Index

faddr

USB Device Functional Address

fifo0

USB FIFO Endpoint 0

fifo1

USB FIFO Endpoint 1

fifo2

USB FIFO Endpoint 2

fifo3

USB FIFO Endpoint 3

fifo4

USB FIFO Endpoint 4

fifo5

USB FIFO Endpoint 5

fifo6

USB FIFO Endpoint 6

fifo7

USB FIFO Endpoint 7

frame

USB Frame Value

fseof

USB Full-Speed Last Transaction to End of Frame Timing

gpcs

USB General-Purpose Control and Status

idvim

USB ID Valid Detect Interrupt Mask

idvisc

USB ID Valid Detect Interrupt Status and Clear

idvris

USB ID Valid Detect Raw Interrupt Status

ie

USB Interrupt Enable

is

USB General Interrupt Status

lseof

USB Low-Speed Last Transaction to End of Frame Timing

naklmt

USB NAK Limit

power

USB Power

pp

USB Peripheral Properties

rqpktcount1

USB Request Packet Count in Block Transfer Endpoint 1

rqpktcount2

USB Request Packet Count in Block Transfer Endpoint 2

rqpktcount3

USB Request Packet Count in Block Transfer Endpoint 3

rqpktcount4

USB Request Packet Count in Block Transfer Endpoint 4

rqpktcount5

USB Request Packet Count in Block Transfer Endpoint 5

rqpktcount6

USB Request Packet Count in Block Transfer Endpoint 6

rqpktcount7

USB Request Packet Count in Block Transfer Endpoint 7

rxcount1

USB Receive Byte Count Endpoint 1

rxcount2

USB Receive Byte Count Endpoint 2

rxcount3

USB Receive Byte Count Endpoint 3

rxcount4

USB Receive Byte Count Endpoint 4

rxcount5

USB Receive Byte Count Endpoint 5

rxcount6

USB Receive Byte Count Endpoint 6

rxcount7

USB Receive Byte Count Endpoint 7

rxcsrh1

USB Receive Control and Status Endpoint 1 High

rxcsrh2

USB Receive Control and Status Endpoint 2 High

rxcsrh3

USB Receive Control and Status Endpoint 3 High

rxcsrh4

USB Receive Control and Status Endpoint 4 High

rxcsrh5

USB Receive Control and Status Endpoint 5 High

rxcsrh6

USB Receive Control and Status Endpoint 6 High

rxcsrh7

USB Receive Control and Status Endpoint 7 High

rxcsrl1

USB Receive Control and Status Endpoint 1 Low

rxcsrl2

USB Receive Control and Status Endpoint 2 Low

rxcsrl3

USB Receive Control and Status Endpoint 3 Low

rxcsrl4

USB Receive Control and Status Endpoint 4 Low

rxcsrl5

USB Receive Control and Status Endpoint 5 Low

rxcsrl6

USB Receive Control and Status Endpoint 6 Low

rxcsrl7

USB Receive Control and Status Endpoint 7 Low

rxdpktbufdis

USB Receive Double Packet Buffer Disable

rxfifoadd

USB Receive FIFO Start Address

rxfifosz

USB Receive Dynamic FIFO Sizing

rxfuncaddr1

USB Receive Functional Address Endpoint 1

rxfuncaddr2

USB Receive Functional Address Endpoint 2

rxfuncaddr3

USB Receive Functional Address Endpoint 3

rxfuncaddr4

USB Receive Functional Address Endpoint 4

rxfuncaddr5

USB Receive Functional Address Endpoint 5

rxfuncaddr6

USB Receive Functional Address Endpoint 6

rxfuncaddr7

USB Receive Functional Address Endpoint 7

rxhubaddr1

USB Receive Hub Address Endpoint 1

rxhubaddr2

USB Receive Hub Address Endpoint 2

rxhubaddr3

USB Receive Hub Address Endpoint 3

rxhubaddr4

USB Receive Hub Address Endpoint 4

rxhubaddr5

USB Receive Hub Address Endpoint 5

rxhubaddr6

USB Receive Hub Address Endpoint 6

rxhubaddr7

USB Receive Hub Address Endpoint 7

rxhubport1

USB Receive Hub Port Endpoint 1

rxhubport2

USB Receive Hub Port Endpoint 2

rxhubport3

USB Receive Hub Port Endpoint 3

rxhubport4

USB Receive Hub Port Endpoint 4

rxhubport5

USB Receive Hub Port Endpoint 5

rxhubport6

USB Receive Hub Port Endpoint 6

rxhubport7

USB Receive Hub Port Endpoint 7

rxie

USB Receive Interrupt Enable

rxinterval1

USB Host Receive Polling Interval Endpoint 1

rxinterval2

USB Host Receive Polling Interval Endpoint 2

rxinterval3

USB Host Receive Polling Interval Endpoint 3

rxinterval4

USB Host Receive Polling Interval Endpoint 4

rxinterval5

USB Host Receive Polling Interval Endpoint 5

rxinterval6

USB Host Receive Polling Interval Endpoint 6

rxinterval7

USB Host Receive Polling Interval Endpoint 7

rxis

USB Receive Interrupt Status

rxmaxp1

USB Maximum Receive Data Endpoint 1

rxmaxp2

USB Maximum Receive Data Endpoint 2

rxmaxp3

USB Maximum Receive Data Endpoint 3

rxmaxp4

USB Maximum Receive Data Endpoint 4

rxmaxp5

USB Maximum Receive Data Endpoint 5

rxmaxp6

USB Maximum Receive Data Endpoint 6

rxmaxp7

USB Maximum Receive Data Endpoint 7

rxtype1

USB Host Configure Receive Type Endpoint 1

rxtype2

USB Host Configure Receive Type Endpoint 2

rxtype3

USB Host Configure Receive Type Endpoint 3

rxtype4

USB Host Configure Receive Type Endpoint 4

rxtype5

USB Host Configure Receive Type Endpoint 5

rxtype6

USB Host Configure Receive Type Endpoint 6

rxtype7

USB Host Configure Receive Type Endpoint 7

test

USB Test Mode

txcsrh1

USB Transmit Control and Status Endpoint 1 High

txcsrh2

USB Transmit Control and Status Endpoint 2 High

txcsrh3

USB Transmit Control and Status Endpoint 3 High

txcsrh4

USB Transmit Control and Status Endpoint 4 High

txcsrh5

USB Transmit Control and Status Endpoint 5 High

txcsrh6

USB Transmit Control and Status Endpoint 6 High

txcsrh7

USB Transmit Control and Status Endpoint 7 High

txcsrl1

USB Transmit Control and Status Endpoint 1 Low

txcsrl2

USB Transmit Control and Status Endpoint 2 Low

txcsrl3

USB Transmit Control and Status Endpoint 3 Low

txcsrl4

USB Transmit Control and Status Endpoint 4 Low

txcsrl5

USB Transmit Control and Status Endpoint 5 Low

txcsrl6

USB Transmit Control and Status Endpoint 6 Low

txcsrl7

USB Transmit Control and Status Endpoint 7 Low

txdpktbufdis

USB Transmit Double Packet Buffer Disable

txfifoadd

USB Transmit FIFO Start Address

txfifosz

USB Transmit Dynamic FIFO Sizing

txfuncaddr0

USB Transmit Functional Address Endpoint 0

txfuncaddr1

USB Transmit Functional Address Endpoint 1

txfuncaddr2

USB Transmit Functional Address Endpoint 2

txfuncaddr3

USB Transmit Functional Address Endpoint 3

txfuncaddr4

USB Transmit Functional Address Endpoint 4

txfuncaddr5

USB Transmit Functional Address Endpoint 5

txfuncaddr6

USB Transmit Functional Address Endpoint 6

txfuncaddr7

USB Transmit Functional Address Endpoint 7

txhubaddr0

USB Transmit Hub Address Endpoint 0

txhubaddr1

USB Transmit Hub Address Endpoint 1

txhubaddr2

USB Transmit Hub Address Endpoint 2

txhubaddr3

USB Transmit Hub Address Endpoint 3

txhubaddr4

USB Transmit Hub Address Endpoint 4

txhubaddr5

USB Transmit Hub Address Endpoint 5

txhubaddr6

USB Transmit Hub Address Endpoint 6

txhubaddr7

USB Transmit Hub Address Endpoint 7

txhubport0

USB Transmit Hub Port Endpoint 0

txhubport1

USB Transmit Hub Port Endpoint 1

txhubport2

USB Transmit Hub Port Endpoint 2

txhubport3

USB Transmit Hub Port Endpoint 3

txhubport4

USB Transmit Hub Port Endpoint 4

txhubport5

USB Transmit Hub Port Endpoint 5

txhubport6

USB Transmit Hub Port Endpoint 6

txhubport7

USB Transmit Hub Port Endpoint 7

txie

USB Transmit Interrupt Enable

txinterval1

USB Host Transmit Interval Endpoint 1

txinterval2

USB Host Transmit Interval Endpoint 2

txinterval3

USB Host Transmit Interval Endpoint 3

txinterval4

USB Host Transmit Interval Endpoint 4

txinterval5

USB Host Transmit Interval Endpoint 5

txinterval6

USB Host Transmit Interval Endpoint 6

txinterval7

USB Host Transmit Interval Endpoint 7

txis

USB Transmit Interrupt Status

txmaxp1

USB Maximum Transmit Data Endpoint 1

txmaxp2

USB Maximum Transmit Data Endpoint 2

txmaxp3

USB Maximum Transmit Data Endpoint 3

txmaxp4

USB Maximum Transmit Data Endpoint 4

txmaxp5

USB Maximum Transmit Data Endpoint 5

txmaxp6

USB Maximum Transmit Data Endpoint 6

txmaxp7

USB Maximum Transmit Data Endpoint 7

txtype1

USB Host Transmit Configure Type Endpoint 1

txtype2

USB Host Transmit Configure Type Endpoint 2

txtype3

USB Host Transmit Configure Type Endpoint 3

txtype4

USB Host Transmit Configure Type Endpoint 4

txtype5

USB Host Transmit Configure Type Endpoint 5

txtype6

USB Host Transmit Configure Type Endpoint 6

txtype7

USB Host Transmit Configure Type Endpoint 7

type0

USB Type Endpoint 0

vdc

USB VBUS Droop Control

vdcim

USB VBUS Droop Control Interrupt Mask

vdcisc

USB VBUS Droop Control Interrupt Status and Clear

vdcris

USB VBUS Droop Control Raw Interrupt Status

vplen

USB OTG VBUS Pulse Timing

Structs

RegisterBlock

Register block

Type Definitions

CONTIM

USB Connect Timing

COUNT0

USB Receive Byte Count Endpoint 0

CSRH0

USB Control and Status Endpoint 0 High

CSRL0

USB Control and Status Endpoint 0 Low

DEVCTL

USB Device Control

DMASEL

USB DMA Select

DRIM

USB Device RESUME Interrupt Mask

DRISC

USB Device RESUME Interrupt Status and Clear

DRRIS

USB Device RESUME Raw Interrupt Status

EPC

USB External Power Control

EPCIM

USB External Power Control Interrupt Mask

EPCISC

USB External Power Control Interrupt Status and Clear

EPCRIS

USB External Power Control Raw Interrupt Status

EPIDX

USB Endpoint Index

FADDR

USB Device Functional Address

FIFO0

USB FIFO Endpoint 0

FIFO1

USB FIFO Endpoint 1

FIFO2

USB FIFO Endpoint 2

FIFO3

USB FIFO Endpoint 3

FIFO4

USB FIFO Endpoint 4

FIFO5

USB FIFO Endpoint 5

FIFO6

USB FIFO Endpoint 6

FIFO7

USB FIFO Endpoint 7

FRAME

USB Frame Value

FSEOF

USB Full-Speed Last Transaction to End of Frame Timing

GPCS

USB General-Purpose Control and Status

IDVIM

USB ID Valid Detect Interrupt Mask

IDVISC

USB ID Valid Detect Interrupt Status and Clear

IDVRIS

USB ID Valid Detect Raw Interrupt Status

IE

USB Interrupt Enable

IS

USB General Interrupt Status

LSEOF

USB Low-Speed Last Transaction to End of Frame Timing

NAKLMT

USB NAK Limit

POWER

USB Power

PP

USB Peripheral Properties

RQPKTCOUNT1

USB Request Packet Count in Block Transfer Endpoint 1

RQPKTCOUNT2

USB Request Packet Count in Block Transfer Endpoint 2

RQPKTCOUNT3

USB Request Packet Count in Block Transfer Endpoint 3

RQPKTCOUNT4

USB Request Packet Count in Block Transfer Endpoint 4

RQPKTCOUNT5

USB Request Packet Count in Block Transfer Endpoint 5

RQPKTCOUNT6

USB Request Packet Count in Block Transfer Endpoint 6

RQPKTCOUNT7

USB Request Packet Count in Block Transfer Endpoint 7

RXCOUNT1

USB Receive Byte Count Endpoint 1

RXCOUNT2

USB Receive Byte Count Endpoint 2

RXCOUNT3

USB Receive Byte Count Endpoint 3

RXCOUNT4

USB Receive Byte Count Endpoint 4

RXCOUNT5

USB Receive Byte Count Endpoint 5

RXCOUNT6

USB Receive Byte Count Endpoint 6

RXCOUNT7

USB Receive Byte Count Endpoint 7

RXCSRH1

USB Receive Control and Status Endpoint 1 High

RXCSRH2

USB Receive Control and Status Endpoint 2 High

RXCSRH3

USB Receive Control and Status Endpoint 3 High

RXCSRH4

USB Receive Control and Status Endpoint 4 High

RXCSRH5

USB Receive Control and Status Endpoint 5 High

RXCSRH6

USB Receive Control and Status Endpoint 6 High

RXCSRH7

USB Receive Control and Status Endpoint 7 High

RXCSRL1

USB Receive Control and Status Endpoint 1 Low

RXCSRL2

USB Receive Control and Status Endpoint 2 Low

RXCSRL3

USB Receive Control and Status Endpoint 3 Low

RXCSRL4

USB Receive Control and Status Endpoint 4 Low

RXCSRL5

USB Receive Control and Status Endpoint 5 Low

RXCSRL6

USB Receive Control and Status Endpoint 6 Low

RXCSRL7

USB Receive Control and Status Endpoint 7 Low

RXDPKTBUFDIS

USB Receive Double Packet Buffer Disable

RXFIFOADD

USB Receive FIFO Start Address

RXFIFOSZ

USB Receive Dynamic FIFO Sizing

RXFUNCADDR1

USB Receive Functional Address Endpoint 1

RXFUNCADDR2

USB Receive Functional Address Endpoint 2

RXFUNCADDR3

USB Receive Functional Address Endpoint 3

RXFUNCADDR4

USB Receive Functional Address Endpoint 4

RXFUNCADDR5

USB Receive Functional Address Endpoint 5

RXFUNCADDR6

USB Receive Functional Address Endpoint 6

RXFUNCADDR7

USB Receive Functional Address Endpoint 7

RXHUBADDR1

USB Receive Hub Address Endpoint 1

RXHUBADDR2

USB Receive Hub Address Endpoint 2

RXHUBADDR3

USB Receive Hub Address Endpoint 3

RXHUBADDR4

USB Receive Hub Address Endpoint 4

RXHUBADDR5

USB Receive Hub Address Endpoint 5

RXHUBADDR6

USB Receive Hub Address Endpoint 6

RXHUBADDR7

USB Receive Hub Address Endpoint 7

RXHUBPORT1

USB Receive Hub Port Endpoint 1

RXHUBPORT2

USB Receive Hub Port Endpoint 2

RXHUBPORT3

USB Receive Hub Port Endpoint 3

RXHUBPORT4

USB Receive Hub Port Endpoint 4

RXHUBPORT5

USB Receive Hub Port Endpoint 5

RXHUBPORT6

USB Receive Hub Port Endpoint 6

RXHUBPORT7

USB Receive Hub Port Endpoint 7

RXIE

USB Receive Interrupt Enable

RXINTERVAL1

USB Host Receive Polling Interval Endpoint 1

RXINTERVAL2

USB Host Receive Polling Interval Endpoint 2

RXINTERVAL3

USB Host Receive Polling Interval Endpoint 3

RXINTERVAL4

USB Host Receive Polling Interval Endpoint 4

RXINTERVAL5

USB Host Receive Polling Interval Endpoint 5

RXINTERVAL6

USB Host Receive Polling Interval Endpoint 6

RXINTERVAL7

USB Host Receive Polling Interval Endpoint 7

RXIS

USB Receive Interrupt Status

RXMAXP1

USB Maximum Receive Data Endpoint 1

RXMAXP2

USB Maximum Receive Data Endpoint 2

RXMAXP3

USB Maximum Receive Data Endpoint 3

RXMAXP4

USB Maximum Receive Data Endpoint 4

RXMAXP5

USB Maximum Receive Data Endpoint 5

RXMAXP6

USB Maximum Receive Data Endpoint 6

RXMAXP7

USB Maximum Receive Data Endpoint 7

RXTYPE1

USB Host Configure Receive Type Endpoint 1

RXTYPE2

USB Host Configure Receive Type Endpoint 2

RXTYPE3

USB Host Configure Receive Type Endpoint 3

RXTYPE4

USB Host Configure Receive Type Endpoint 4

RXTYPE5

USB Host Configure Receive Type Endpoint 5

RXTYPE6

USB Host Configure Receive Type Endpoint 6

RXTYPE7

USB Host Configure Receive Type Endpoint 7

TEST

USB Test Mode

TXCSRH1

USB Transmit Control and Status Endpoint 1 High

TXCSRH2

USB Transmit Control and Status Endpoint 2 High

TXCSRH3

USB Transmit Control and Status Endpoint 3 High

TXCSRH4

USB Transmit Control and Status Endpoint 4 High

TXCSRH5

USB Transmit Control and Status Endpoint 5 High

TXCSRH6

USB Transmit Control and Status Endpoint 6 High

TXCSRH7

USB Transmit Control and Status Endpoint 7 High

TXCSRL1

USB Transmit Control and Status Endpoint 1 Low

TXCSRL2

USB Transmit Control and Status Endpoint 2 Low

TXCSRL3

USB Transmit Control and Status Endpoint 3 Low

TXCSRL4

USB Transmit Control and Status Endpoint 4 Low

TXCSRL5

USB Transmit Control and Status Endpoint 5 Low

TXCSRL6

USB Transmit Control and Status Endpoint 6 Low

TXCSRL7

USB Transmit Control and Status Endpoint 7 Low

TXDPKTBUFDIS

USB Transmit Double Packet Buffer Disable

TXFIFOADD

USB Transmit FIFO Start Address

TXFIFOSZ

USB Transmit Dynamic FIFO Sizing

TXFUNCADDR0

USB Transmit Functional Address Endpoint 0

TXFUNCADDR1

USB Transmit Functional Address Endpoint 1

TXFUNCADDR2

USB Transmit Functional Address Endpoint 2

TXFUNCADDR3

USB Transmit Functional Address Endpoint 3

TXFUNCADDR4

USB Transmit Functional Address Endpoint 4

TXFUNCADDR5

USB Transmit Functional Address Endpoint 5

TXFUNCADDR6

USB Transmit Functional Address Endpoint 6

TXFUNCADDR7

USB Transmit Functional Address Endpoint 7

TXHUBADDR0

USB Transmit Hub Address Endpoint 0

TXHUBADDR1

USB Transmit Hub Address Endpoint 1

TXHUBADDR2

USB Transmit Hub Address Endpoint 2

TXHUBADDR3

USB Transmit Hub Address Endpoint 3

TXHUBADDR4

USB Transmit Hub Address Endpoint 4

TXHUBADDR5

USB Transmit Hub Address Endpoint 5

TXHUBADDR6

USB Transmit Hub Address Endpoint 6

TXHUBADDR7

USB Transmit Hub Address Endpoint 7

TXHUBPORT0

USB Transmit Hub Port Endpoint 0

TXHUBPORT1

USB Transmit Hub Port Endpoint 1

TXHUBPORT2

USB Transmit Hub Port Endpoint 2

TXHUBPORT3

USB Transmit Hub Port Endpoint 3

TXHUBPORT4

USB Transmit Hub Port Endpoint 4

TXHUBPORT5

USB Transmit Hub Port Endpoint 5

TXHUBPORT6

USB Transmit Hub Port Endpoint 6

TXHUBPORT7

USB Transmit Hub Port Endpoint 7

TXIE

USB Transmit Interrupt Enable

TXINTERVAL1

USB Host Transmit Interval Endpoint 1

TXINTERVAL2

USB Host Transmit Interval Endpoint 2

TXINTERVAL3

USB Host Transmit Interval Endpoint 3

TXINTERVAL4

USB Host Transmit Interval Endpoint 4

TXINTERVAL5

USB Host Transmit Interval Endpoint 5

TXINTERVAL6

USB Host Transmit Interval Endpoint 6

TXINTERVAL7

USB Host Transmit Interval Endpoint 7

TXIS

USB Transmit Interrupt Status

TXMAXP1

USB Maximum Transmit Data Endpoint 1

TXMAXP2

USB Maximum Transmit Data Endpoint 2

TXMAXP3

USB Maximum Transmit Data Endpoint 3

TXMAXP4

USB Maximum Transmit Data Endpoint 4

TXMAXP5

USB Maximum Transmit Data Endpoint 5

TXMAXP6

USB Maximum Transmit Data Endpoint 6

TXMAXP7

USB Maximum Transmit Data Endpoint 7

TXTYPE1

USB Host Transmit Configure Type Endpoint 1

TXTYPE2

USB Host Transmit Configure Type Endpoint 2

TXTYPE3

USB Host Transmit Configure Type Endpoint 3

TXTYPE4

USB Host Transmit Configure Type Endpoint 4

TXTYPE5

USB Host Transmit Configure Type Endpoint 5

TXTYPE6

USB Host Transmit Configure Type Endpoint 6

TXTYPE7

USB Host Transmit Configure Type Endpoint 7

TYPE0

USB Type Endpoint 0

VDC

USB VBUS Droop Control

VDCIM

USB VBUS Droop Control Interrupt Mask

VDCISC

USB VBUS Droop Control Interrupt Status and Clear

VDCRIS

USB VBUS Droop Control Raw Interrupt Status

VPLEN

USB OTG VBUS Pulse Timing