Struct tm4c123x::usb0::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub faddr: FADDR,
    pub power: POWER,
    pub txis: TXIS,
    pub rxis: RXIS,
    pub txie: TXIE,
    pub rxie: RXIE,
    pub is: IS,
    pub ie: IE,
    pub frame: FRAME,
    pub epidx: EPIDX,
    pub test: TEST,
    pub fifo0: FIFO0,
    pub fifo1: FIFO1,
    pub fifo2: FIFO2,
    pub fifo3: FIFO3,
    pub fifo4: FIFO4,
    pub fifo5: FIFO5,
    pub fifo6: FIFO6,
    pub fifo7: FIFO7,
    pub devctl: DEVCTL,
    pub txfifosz: TXFIFOSZ,
    pub rxfifosz: RXFIFOSZ,
    pub txfifoadd: TXFIFOADD,
    pub rxfifoadd: RXFIFOADD,
    pub contim: CONTIM,
    pub vplen: VPLEN,
    pub fseof: FSEOF,
    pub lseof: LSEOF,
    pub txfuncaddr0: TXFUNCADDR0,
    pub txhubaddr0: TXHUBADDR0,
    pub txhubport0: TXHUBPORT0,
    pub txfuncaddr1: TXFUNCADDR1,
    pub txhubaddr1: TXHUBADDR1,
    pub txhubport1: TXHUBPORT1,
    pub rxfuncaddr1: RXFUNCADDR1,
    pub rxhubaddr1: RXHUBADDR1,
    pub rxhubport1: RXHUBPORT1,
    pub txfuncaddr2: TXFUNCADDR2,
    pub txhubaddr2: TXHUBADDR2,
    pub txhubport2: TXHUBPORT2,
    pub rxfuncaddr2: RXFUNCADDR2,
    pub rxhubaddr2: RXHUBADDR2,
    pub rxhubport2: RXHUBPORT2,
    pub txfuncaddr3: TXFUNCADDR3,
    pub txhubaddr3: TXHUBADDR3,
    pub txhubport3: TXHUBPORT3,
    pub rxfuncaddr3: RXFUNCADDR3,
    pub rxhubaddr3: RXHUBADDR3,
    pub rxhubport3: RXHUBPORT3,
    pub txfuncaddr4: TXFUNCADDR4,
    pub txhubaddr4: TXHUBADDR4,
    pub txhubport4: TXHUBPORT4,
    pub rxfuncaddr4: RXFUNCADDR4,
    pub rxhubaddr4: RXHUBADDR4,
    pub rxhubport4: RXHUBPORT4,
    pub txfuncaddr5: TXFUNCADDR5,
    pub txhubaddr5: TXHUBADDR5,
    pub txhubport5: TXHUBPORT5,
    pub rxfuncaddr5: RXFUNCADDR5,
    pub rxhubaddr5: RXHUBADDR5,
    pub rxhubport5: RXHUBPORT5,
    pub txfuncaddr6: TXFUNCADDR6,
    pub txhubaddr6: TXHUBADDR6,
    pub txhubport6: TXHUBPORT6,
    pub rxfuncaddr6: RXFUNCADDR6,
    pub rxhubaddr6: RXHUBADDR6,
    pub rxhubport6: RXHUBPORT6,
    pub txfuncaddr7: TXFUNCADDR7,
    pub txhubaddr7: TXHUBADDR7,
    pub txhubport7: TXHUBPORT7,
    pub rxfuncaddr7: RXFUNCADDR7,
    pub rxhubaddr7: RXHUBADDR7,
    pub rxhubport7: RXHUBPORT7,
    pub csrl0: CSRL0,
    pub csrh0: CSRH0,
    pub count0: COUNT0,
    pub type0: TYPE0,
    pub naklmt: NAKLMT,
    pub txmaxp1: TXMAXP1,
    pub txcsrl1: TXCSRL1,
    pub txcsrh1: TXCSRH1,
    pub rxmaxp1: RXMAXP1,
    pub rxcsrl1: RXCSRL1,
    pub rxcsrh1: RXCSRH1,
    pub rxcount1: RXCOUNT1,
    pub txtype1: TXTYPE1,
    pub txinterval1: TXINTERVAL1,
    pub rxtype1: RXTYPE1,
    pub rxinterval1: RXINTERVAL1,
    pub txmaxp2: TXMAXP2,
    pub txcsrl2: TXCSRL2,
    pub txcsrh2: TXCSRH2,
    pub rxmaxp2: RXMAXP2,
    pub rxcsrl2: RXCSRL2,
    pub rxcsrh2: RXCSRH2,
    pub rxcount2: RXCOUNT2,
    pub txtype2: TXTYPE2,
    pub txinterval2: TXINTERVAL2,
    pub rxtype2: RXTYPE2,
    pub rxinterval2: RXINTERVAL2,
    pub txmaxp3: TXMAXP3,
    pub txcsrl3: TXCSRL3,
    pub txcsrh3: TXCSRH3,
    pub rxmaxp3: RXMAXP3,
    pub rxcsrl3: RXCSRL3,
    pub rxcsrh3: RXCSRH3,
    pub rxcount3: RXCOUNT3,
    pub txtype3: TXTYPE3,
    pub txinterval3: TXINTERVAL3,
    pub rxtype3: RXTYPE3,
    pub rxinterval3: RXINTERVAL3,
    pub txmaxp4: TXMAXP4,
    pub txcsrl4: TXCSRL4,
    pub txcsrh4: TXCSRH4,
    pub rxmaxp4: RXMAXP4,
    pub rxcsrl4: RXCSRL4,
    pub rxcsrh4: RXCSRH4,
    pub rxcount4: RXCOUNT4,
    pub txtype4: TXTYPE4,
    pub txinterval4: TXINTERVAL4,
    pub rxtype4: RXTYPE4,
    pub rxinterval4: RXINTERVAL4,
    pub txmaxp5: TXMAXP5,
    pub txcsrl5: TXCSRL5,
    pub txcsrh5: TXCSRH5,
    pub rxmaxp5: RXMAXP5,
    pub rxcsrl5: RXCSRL5,
    pub rxcsrh5: RXCSRH5,
    pub rxcount5: RXCOUNT5,
    pub txtype5: TXTYPE5,
    pub txinterval5: TXINTERVAL5,
    pub rxtype5: RXTYPE5,
    pub rxinterval5: RXINTERVAL5,
    pub txmaxp6: TXMAXP6,
    pub txcsrl6: TXCSRL6,
    pub txcsrh6: TXCSRH6,
    pub rxmaxp6: RXMAXP6,
    pub rxcsrl6: RXCSRL6,
    pub rxcsrh6: RXCSRH6,
    pub rxcount6: RXCOUNT6,
    pub txtype6: TXTYPE6,
    pub txinterval6: TXINTERVAL6,
    pub rxtype6: RXTYPE6,
    pub rxinterval6: RXINTERVAL6,
    pub txmaxp7: TXMAXP7,
    pub txcsrl7: TXCSRL7,
    pub txcsrh7: TXCSRH7,
    pub rxmaxp7: RXMAXP7,
    pub rxcsrl7: RXCSRL7,
    pub rxcsrh7: RXCSRH7,
    pub rxcount7: RXCOUNT7,
    pub txtype7: TXTYPE7,
    pub txinterval7: TXINTERVAL7,
    pub rxtype7: RXTYPE7,
    pub rxinterval7: RXINTERVAL7,
    pub rqpktcount1: RQPKTCOUNT1,
    pub rqpktcount2: RQPKTCOUNT2,
    pub rqpktcount3: RQPKTCOUNT3,
    pub rqpktcount4: RQPKTCOUNT4,
    pub rqpktcount5: RQPKTCOUNT5,
    pub rqpktcount6: RQPKTCOUNT6,
    pub rqpktcount7: RQPKTCOUNT7,
    pub rxdpktbufdis: RXDPKTBUFDIS,
    pub txdpktbufdis: TXDPKTBUFDIS,
    pub epc: EPC,
    pub epcris: EPCRIS,
    pub epcim: EPCIM,
    pub epcisc: EPCISC,
    pub drris: DRRIS,
    pub drim: DRIM,
    pub drisc: DRISC,
    pub gpcs: GPCS,
    pub vdc: VDC,
    pub vdcris: VDCRIS,
    pub vdcim: VDCIM,
    pub vdcisc: VDCISC,
    pub idvris: IDVRIS,
    pub idvim: IDVIM,
    pub idvisc: IDVISC,
    pub dmasel: DMASEL,
    pub pp: PP,
    /* private fields */

}

Register block

Fields
faddr: FADDR

0x00 - USB Device Functional Address

power: POWER

0x01 - USB Power

txis: TXIS

0x02 - USB Transmit Interrupt Status

rxis: RXIS

0x04 - USB Receive Interrupt Status

txie: TXIE

0x06 - USB Transmit Interrupt Enable

rxie: RXIE

0x08 - USB Receive Interrupt Enable

is: IS

0x0a - USB General Interrupt Status

ie: IE

0x0b - USB Interrupt Enable

frame: FRAME

0x0c - USB Frame Value

epidx: EPIDX

0x0e - USB Endpoint Index

test: TEST

0x0f - USB Test Mode

fifo0: FIFO0

0x20 - USB FIFO Endpoint 0

fifo1: FIFO1

0x24 - USB FIFO Endpoint 1

fifo2: FIFO2

0x28 - USB FIFO Endpoint 2

fifo3: FIFO3

0x2c - USB FIFO Endpoint 3

fifo4: FIFO4

0x30 - USB FIFO Endpoint 4

fifo5: FIFO5

0x34 - USB FIFO Endpoint 5

fifo6: FIFO6

0x38 - USB FIFO Endpoint 6

fifo7: FIFO7

0x3c - USB FIFO Endpoint 7

devctl: DEVCTL

0x60 - USB Device Control

txfifosz: TXFIFOSZ

0x62 - USB Transmit Dynamic FIFO Sizing

rxfifosz: RXFIFOSZ

0x63 - USB Receive Dynamic FIFO Sizing

txfifoadd: TXFIFOADD

0x64 - USB Transmit FIFO Start Address

rxfifoadd: RXFIFOADD

0x66 - USB Receive FIFO Start Address

contim: CONTIM

0x7a - USB Connect Timing

vplen: VPLEN

0x7b - USB OTG VBUS Pulse Timing

fseof: FSEOF

0x7d - USB Full-Speed Last Transaction to End of Frame Timing

lseof: LSEOF

0x7e - USB Low-Speed Last Transaction to End of Frame Timing

txfuncaddr0: TXFUNCADDR0

0x80 - USB Transmit Functional Address Endpoint 0

txhubaddr0: TXHUBADDR0

0x82 - USB Transmit Hub Address Endpoint 0

txhubport0: TXHUBPORT0

0x83 - USB Transmit Hub Port Endpoint 0

txfuncaddr1: TXFUNCADDR1

0x88 - USB Transmit Functional Address Endpoint 1

txhubaddr1: TXHUBADDR1

0x8a - USB Transmit Hub Address Endpoint 1

txhubport1: TXHUBPORT1

0x8b - USB Transmit Hub Port Endpoint 1

rxfuncaddr1: RXFUNCADDR1

0x8c - USB Receive Functional Address Endpoint 1

rxhubaddr1: RXHUBADDR1

0x8e - USB Receive Hub Address Endpoint 1

rxhubport1: RXHUBPORT1

0x8f - USB Receive Hub Port Endpoint 1

txfuncaddr2: TXFUNCADDR2

0x90 - USB Transmit Functional Address Endpoint 2

txhubaddr2: TXHUBADDR2

0x92 - USB Transmit Hub Address Endpoint 2

txhubport2: TXHUBPORT2

0x93 - USB Transmit Hub Port Endpoint 2

rxfuncaddr2: RXFUNCADDR2

0x94 - USB Receive Functional Address Endpoint 2

rxhubaddr2: RXHUBADDR2

0x96 - USB Receive Hub Address Endpoint 2

rxhubport2: RXHUBPORT2

0x97 - USB Receive Hub Port Endpoint 2

txfuncaddr3: TXFUNCADDR3

0x98 - USB Transmit Functional Address Endpoint 3

txhubaddr3: TXHUBADDR3

0x9a - USB Transmit Hub Address Endpoint 3

txhubport3: TXHUBPORT3

0x9b - USB Transmit Hub Port Endpoint 3

rxfuncaddr3: RXFUNCADDR3

0x9c - USB Receive Functional Address Endpoint 3

rxhubaddr3: RXHUBADDR3

0x9e - USB Receive Hub Address Endpoint 3

rxhubport3: RXHUBPORT3

0x9f - USB Receive Hub Port Endpoint 3

txfuncaddr4: TXFUNCADDR4

0xa0 - USB Transmit Functional Address Endpoint 4

txhubaddr4: TXHUBADDR4

0xa2 - USB Transmit Hub Address Endpoint 4

txhubport4: TXHUBPORT4

0xa3 - USB Transmit Hub Port Endpoint 4

rxfuncaddr4: RXFUNCADDR4

0xa4 - USB Receive Functional Address Endpoint 4

rxhubaddr4: RXHUBADDR4

0xa6 - USB Receive Hub Address Endpoint 4

rxhubport4: RXHUBPORT4

0xa7 - USB Receive Hub Port Endpoint 4

txfuncaddr5: TXFUNCADDR5

0xa8 - USB Transmit Functional Address Endpoint 5

txhubaddr5: TXHUBADDR5

0xaa - USB Transmit Hub Address Endpoint 5

txhubport5: TXHUBPORT5

0xab - USB Transmit Hub Port Endpoint 5

rxfuncaddr5: RXFUNCADDR5

0xac - USB Receive Functional Address Endpoint 5

rxhubaddr5: RXHUBADDR5

0xae - USB Receive Hub Address Endpoint 5

rxhubport5: RXHUBPORT5

0xaf - USB Receive Hub Port Endpoint 5

txfuncaddr6: TXFUNCADDR6

0xb0 - USB Transmit Functional Address Endpoint 6

txhubaddr6: TXHUBADDR6

0xb2 - USB Transmit Hub Address Endpoint 6

txhubport6: TXHUBPORT6

0xb3 - USB Transmit Hub Port Endpoint 6

rxfuncaddr6: RXFUNCADDR6

0xb4 - USB Receive Functional Address Endpoint 6

rxhubaddr6: RXHUBADDR6

0xb6 - USB Receive Hub Address Endpoint 6

rxhubport6: RXHUBPORT6

0xb7 - USB Receive Hub Port Endpoint 6

txfuncaddr7: TXFUNCADDR7

0xb8 - USB Transmit Functional Address Endpoint 7

txhubaddr7: TXHUBADDR7

0xba - USB Transmit Hub Address Endpoint 7

txhubport7: TXHUBPORT7

0xbb - USB Transmit Hub Port Endpoint 7

rxfuncaddr7: RXFUNCADDR7

0xbc - USB Receive Functional Address Endpoint 7

rxhubaddr7: RXHUBADDR7

0xbe - USB Receive Hub Address Endpoint 7

rxhubport7: RXHUBPORT7

0xbf - USB Receive Hub Port Endpoint 7

csrl0: CSRL0

0x102 - USB Control and Status Endpoint 0 Low

csrh0: CSRH0

0x103 - USB Control and Status Endpoint 0 High

count0: COUNT0

0x108 - USB Receive Byte Count Endpoint 0

type0: TYPE0

0x10a - USB Type Endpoint 0

naklmt: NAKLMT

0x10b - USB NAK Limit

txmaxp1: TXMAXP1

0x110 - USB Maximum Transmit Data Endpoint 1

txcsrl1: TXCSRL1

0x112 - USB Transmit Control and Status Endpoint 1 Low

txcsrh1: TXCSRH1

0x113 - USB Transmit Control and Status Endpoint 1 High

rxmaxp1: RXMAXP1

0x114 - USB Maximum Receive Data Endpoint 1

rxcsrl1: RXCSRL1

0x116 - USB Receive Control and Status Endpoint 1 Low

rxcsrh1: RXCSRH1

0x117 - USB Receive Control and Status Endpoint 1 High

rxcount1: RXCOUNT1

0x118 - USB Receive Byte Count Endpoint 1

txtype1: TXTYPE1

0x11a - USB Host Transmit Configure Type Endpoint 1

txinterval1: TXINTERVAL1

0x11b - USB Host Transmit Interval Endpoint 1

rxtype1: RXTYPE1

0x11c - USB Host Configure Receive Type Endpoint 1

rxinterval1: RXINTERVAL1

0x11d - USB Host Receive Polling Interval Endpoint 1

txmaxp2: TXMAXP2

0x120 - USB Maximum Transmit Data Endpoint 2

txcsrl2: TXCSRL2

0x122 - USB Transmit Control and Status Endpoint 2 Low

txcsrh2: TXCSRH2

0x123 - USB Transmit Control and Status Endpoint 2 High

rxmaxp2: RXMAXP2

0x124 - USB Maximum Receive Data Endpoint 2

rxcsrl2: RXCSRL2

0x126 - USB Receive Control and Status Endpoint 2 Low

rxcsrh2: RXCSRH2

0x127 - USB Receive Control and Status Endpoint 2 High

rxcount2: RXCOUNT2

0x128 - USB Receive Byte Count Endpoint 2

txtype2: TXTYPE2

0x12a - USB Host Transmit Configure Type Endpoint 2

txinterval2: TXINTERVAL2

0x12b - USB Host Transmit Interval Endpoint 2

rxtype2: RXTYPE2

0x12c - USB Host Configure Receive Type Endpoint 2

rxinterval2: RXINTERVAL2

0x12d - USB Host Receive Polling Interval Endpoint 2

txmaxp3: TXMAXP3

0x130 - USB Maximum Transmit Data Endpoint 3

txcsrl3: TXCSRL3

0x132 - USB Transmit Control and Status Endpoint 3 Low

txcsrh3: TXCSRH3

0x133 - USB Transmit Control and Status Endpoint 3 High

rxmaxp3: RXMAXP3

0x134 - USB Maximum Receive Data Endpoint 3

rxcsrl3: RXCSRL3

0x136 - USB Receive Control and Status Endpoint 3 Low

rxcsrh3: RXCSRH3

0x137 - USB Receive Control and Status Endpoint 3 High

rxcount3: RXCOUNT3

0x138 - USB Receive Byte Count Endpoint 3

txtype3: TXTYPE3

0x13a - USB Host Transmit Configure Type Endpoint 3

txinterval3: TXINTERVAL3

0x13b - USB Host Transmit Interval Endpoint 3

rxtype3: RXTYPE3

0x13c - USB Host Configure Receive Type Endpoint 3

rxinterval3: RXINTERVAL3

0x13d - USB Host Receive Polling Interval Endpoint 3

txmaxp4: TXMAXP4

0x140 - USB Maximum Transmit Data Endpoint 4

txcsrl4: TXCSRL4

0x142 - USB Transmit Control and Status Endpoint 4 Low

txcsrh4: TXCSRH4

0x143 - USB Transmit Control and Status Endpoint 4 High

rxmaxp4: RXMAXP4

0x144 - USB Maximum Receive Data Endpoint 4

rxcsrl4: RXCSRL4

0x146 - USB Receive Control and Status Endpoint 4 Low

rxcsrh4: RXCSRH4

0x147 - USB Receive Control and Status Endpoint 4 High

rxcount4: RXCOUNT4

0x148 - USB Receive Byte Count Endpoint 4

txtype4: TXTYPE4

0x14a - USB Host Transmit Configure Type Endpoint 4

txinterval4: TXINTERVAL4

0x14b - USB Host Transmit Interval Endpoint 4

rxtype4: RXTYPE4

0x14c - USB Host Configure Receive Type Endpoint 4

rxinterval4: RXINTERVAL4

0x14d - USB Host Receive Polling Interval Endpoint 4

txmaxp5: TXMAXP5

0x150 - USB Maximum Transmit Data Endpoint 5

txcsrl5: TXCSRL5

0x152 - USB Transmit Control and Status Endpoint 5 Low

txcsrh5: TXCSRH5

0x153 - USB Transmit Control and Status Endpoint 5 High

rxmaxp5: RXMAXP5

0x154 - USB Maximum Receive Data Endpoint 5

rxcsrl5: RXCSRL5

0x156 - USB Receive Control and Status Endpoint 5 Low

rxcsrh5: RXCSRH5

0x157 - USB Receive Control and Status Endpoint 5 High

rxcount5: RXCOUNT5

0x158 - USB Receive Byte Count Endpoint 5

txtype5: TXTYPE5

0x15a - USB Host Transmit Configure Type Endpoint 5

txinterval5: TXINTERVAL5

0x15b - USB Host Transmit Interval Endpoint 5

rxtype5: RXTYPE5

0x15c - USB Host Configure Receive Type Endpoint 5

rxinterval5: RXINTERVAL5

0x15d - USB Host Receive Polling Interval Endpoint 5

txmaxp6: TXMAXP6

0x160 - USB Maximum Transmit Data Endpoint 6

txcsrl6: TXCSRL6

0x162 - USB Transmit Control and Status Endpoint 6 Low

txcsrh6: TXCSRH6

0x163 - USB Transmit Control and Status Endpoint 6 High

rxmaxp6: RXMAXP6

0x164 - USB Maximum Receive Data Endpoint 6

rxcsrl6: RXCSRL6

0x166 - USB Receive Control and Status Endpoint 6 Low

rxcsrh6: RXCSRH6

0x167 - USB Receive Control and Status Endpoint 6 High

rxcount6: RXCOUNT6

0x168 - USB Receive Byte Count Endpoint 6

txtype6: TXTYPE6

0x16a - USB Host Transmit Configure Type Endpoint 6

txinterval6: TXINTERVAL6

0x16b - USB Host Transmit Interval Endpoint 6

rxtype6: RXTYPE6

0x16c - USB Host Configure Receive Type Endpoint 6

rxinterval6: RXINTERVAL6

0x16d - USB Host Receive Polling Interval Endpoint 6

txmaxp7: TXMAXP7

0x170 - USB Maximum Transmit Data Endpoint 7

txcsrl7: TXCSRL7

0x172 - USB Transmit Control and Status Endpoint 7 Low

txcsrh7: TXCSRH7

0x173 - USB Transmit Control and Status Endpoint 7 High

rxmaxp7: RXMAXP7

0x174 - USB Maximum Receive Data Endpoint 7

rxcsrl7: RXCSRL7

0x176 - USB Receive Control and Status Endpoint 7 Low

rxcsrh7: RXCSRH7

0x177 - USB Receive Control and Status Endpoint 7 High

rxcount7: RXCOUNT7

0x178 - USB Receive Byte Count Endpoint 7

txtype7: TXTYPE7

0x17a - USB Host Transmit Configure Type Endpoint 7

txinterval7: TXINTERVAL7

0x17b - USB Host Transmit Interval Endpoint 7

rxtype7: RXTYPE7

0x17c - USB Host Configure Receive Type Endpoint 7

rxinterval7: RXINTERVAL7

0x17d - USB Host Receive Polling Interval Endpoint 7

rqpktcount1: RQPKTCOUNT1

0x304 - USB Request Packet Count in Block Transfer Endpoint 1

rqpktcount2: RQPKTCOUNT2

0x308 - USB Request Packet Count in Block Transfer Endpoint 2

rqpktcount3: RQPKTCOUNT3

0x30c - USB Request Packet Count in Block Transfer Endpoint 3

rqpktcount4: RQPKTCOUNT4

0x310 - USB Request Packet Count in Block Transfer Endpoint 4

rqpktcount5: RQPKTCOUNT5

0x314 - USB Request Packet Count in Block Transfer Endpoint 5

rqpktcount6: RQPKTCOUNT6

0x318 - USB Request Packet Count in Block Transfer Endpoint 6

rqpktcount7: RQPKTCOUNT7

0x31c - USB Request Packet Count in Block Transfer Endpoint 7

rxdpktbufdis: RXDPKTBUFDIS

0x340 - USB Receive Double Packet Buffer Disable

txdpktbufdis: TXDPKTBUFDIS

0x342 - USB Transmit Double Packet Buffer Disable

epc: EPC

0x400 - USB External Power Control

epcris: EPCRIS

0x404 - USB External Power Control Raw Interrupt Status

epcim: EPCIM

0x408 - USB External Power Control Interrupt Mask

epcisc: EPCISC

0x40c - USB External Power Control Interrupt Status and Clear

drris: DRRIS

0x410 - USB Device RESUME Raw Interrupt Status

drim: DRIM

0x414 - USB Device RESUME Interrupt Mask

drisc: DRISC

0x418 - USB Device RESUME Interrupt Status and Clear

gpcs: GPCS

0x41c - USB General-Purpose Control and Status

vdc: VDC

0x430 - USB VBUS Droop Control

vdcris: VDCRIS

0x434 - USB VBUS Droop Control Raw Interrupt Status

vdcim: VDCIM

0x438 - USB VBUS Droop Control Interrupt Mask

vdcisc: VDCISC

0x43c - USB VBUS Droop Control Interrupt Status and Clear

idvris: IDVRIS

0x444 - USB ID Valid Detect Raw Interrupt Status

idvim: IDVIM

0x448 - USB ID Valid Detect Interrupt Mask

idvisc: IDVISC

0x44c - USB ID Valid Detect Interrupt Status and Clear

dmasel: DMASEL

0x450 - USB DMA Select

pp: PP

0xfc0 - USB Peripheral Properties